Add implementation of Committee
This commit adds two mix tasks, which are both vital part to
`Committee`.

The first one is `committee.install`, which "installs" Committee, the
installation is two-part:

- Generate a `commit.exs`, this is the actual code that will be run for
git hooks. The functions here map to git hooks name.
- Generate a stub in `.git/hooks` (which are just bash scripts calling
out to a runner.)

Speaking of runner, that's the second task that this commit introduces.
`committee.runner`.

This runner is basically a shim that gets run on every single git hooks,
and it then runs the corresponding function defined in `commit.exs`.

This commit concludes the basic functionality of `Committee` :)