Add ability to return custom message post-hooks
This commit adds two different ways to print custom message post-hook.

The first way is for when hooks are **successful**, if you return a
`{:ok, message}` tuple then the message will be printed to screen.

The second way is for when hooks are **unsuccessful**, in that case then
you can return a `{:halt, reason}` tuple which will also get printed to
screen.

This commit also removes the default message that gets printed when a
hook is ran, it may be unnecessary noise.