Add `Committee.Helpers` for utility functions
This module provides some helper functions when dealing with git
commits, the most notable one being `staged_files/0`, which is an
easy way for you get a list of currently staged files.

The second one is `branch_name/0`, which is a helper function to help
you get the current branch name. This is helpful if you want to enforce
a certain branch naming convention.