Improve Committee installation template
This commit just improves the comments a bit, such as teaching the
end-users how to abort a commit or print a success message, it also adds
`@impl true` so we get the benefit of Elixir's compile time checking for
behaviour.

Also imports `Committee.Helpers` as to show users that they have access
to helper functions in that module.