Add new `staged_files/1` to filter by suffix
This previous version of `staged_files/0` proved a slight bug, because
`mix format` wouldn't be able to run on non-Elixir code.

Clients could mitigate by writing their own filter, but I figured this
could be provided by the `Helpers` module, as it could also potentially
be used for other purposes.

Hence, this commit introduces a new `staged_commit/1` that takes in
either a string or a list of strings that matches on suffix of the
files.