Warn instead of error if `config` is not provided
This commit is an update #13 in which instead of aborting the commit
when the `commit.exs` file is not found, we will now just log you a
warning that you can safely ignore (but will not impact your normal
workflow), and we will also show you options to uninstall Committee.

This is because if someone else on the team has removed `Committee`
without organizing with the team, everyone elses' commit flow might
break, since the config file is now gone.

The second update in this PR is a major refactor, in which it uses
tagged tuple and the with-syntax to elegantly handle error at each step
of the execution.