Add versioning to committee hooks
Right now Committee will skip installation if a file contains the
identifier, `# committee`. However this means that we won't be able to
change the implementation of the runner file.

With this version, we can check the version to see if we should
regenerate the runner file.

With this commit, if the version field is not the same as the current
version, then we will regenerate/update the hooks. If the version field
is the same as the hooks however, we will skip.